refactor to support testing
