+++ Cargo.toml
@@ -21,7 +21,7 @@ bincode = "2.0.1"
 uuid = { version = "1", features = ["js", "v4"] }
 wasm-bindgen = "0.2.100"
 wasm-bindgen-futures = "0.4.50"
-web-sys = { version = "0.3.77", features = [
+web-sys = { version = "0.3.98", features = [
     "console",
     "Response",
     "Request",
@@ -55,7 +55,7 @@ web-sys = { version = "0.3.77", features = [
 reqwest = { version = "0.12.15", features = ["json", "cookies"] }
 serde_json = "1.0.140"
 serde-wasm-bindgen = "0.6.5"
-js-sys = "0.3.77"
+js-sys = "0.3.98"
 ntor = { git = "https://github.com/globe-and-citizen/ntor.git", tag = "0.1.3" }
 serde = { version = "1.0.219", features = ["derive"] }
 wasm-streams = "0.4.2"
