+++ src/types/mod.rs
@@ -1,5 +1,5 @@
 pub mod http_caller;
 pub mod network_state;
 pub mod request;
-pub(crate) mod response;
-pub(crate) mod service_provider;
+pub mod response;
+pub mod service_provider;
